修复删除错误日志文件

+++ shell/share.sh
@@ -307,5 +307,5 @@ detect_termux
 detect_macos
 define_cmd
 
-rm $task_error_log_path
+[ -f $task_error_log_path ] && rm $task_error_log_path
 import_config $1 >> $task_error_log_path 2>&1
